fixed client component
