+++ web/src/routes/+layout.svelte
@@ -49,11 +49,11 @@
           <span class="lg:hidden inline">👇</span>
           <span class="lg:inline hidden">Try Now</span>
         </a>
-		<a class="btn bg-gradient-to-br variant-gradient-secondary-primary" href="/about">
+        <a class="btn bg-gradient-to-br variant-gradient-secondary-primary" href="/about">
           <span class="lg:hidden inline">🦙</span>
           <span class="lg:inline hidden">About LLaMP </span>
         </a>
-		<a
+        <a
           class="btn bg-gradient-to-br variant-gradient-primary-secondary"
           href="https://materialsproject.org"
           target="_blank"
@@ -63,7 +63,7 @@
 
           <span class="lg:inline hidden">Materials Project</span>
         </a>
-        
+
         <a
           class="btn-icon variant-ghost"
           href="https://github.com/chiang-yuan/llamp"
